+++ app/serializers/asyncapi/server/job_serializer.rb
@@ -2,7 +2,7 @@ module Asyncapi
   module Server
     class JobSerializer < ActiveModel::Serializer
 
-      attributes :id, :url
+      attributes :id, :url, :secret
 
     end
   end