+++ db/migrate/20141212064931_add_secret_to_asyncapi_server_job.rb
@@ -1,0 +1,5 @@
+class AddSecretToAsyncapiServerJob < ActiveRecord::Migration
+  def change
+    add_column :asyncapi_server_jobs, :secret, :string
+  end
+end