+++ spec/dummy/db/migrate/20141212065006_add_secret_to_asyncapi_server_job.asyncapi_server.rb
@@ -1,0 +1,6 @@
+# This migration comes from asyncapi_server (originally 20141212064931)
+class AddSecretToAsyncapiServerJob < ActiveRecord::Migration
+  def change
+    add_column :asyncapi_server_jobs, :secret, :string
+  end
+end