+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # TBA
 
 - Fix issue with asyncapi client not seeing request body; post as json
+- Send secret back to the app using `asyncapi-client` for authentication
 
 # 0.0.1